remove guidance from GAIS requestCompletion
